Update to use new query operators
In my rush to get this out, I forgot to have tyler update the query
operators to use the new syntax. This is now updated. This commit also
includes an update to the Rakefile to add the `test` directory to the
path so the test helper file can easily be loaded. It also adds test to
the base query class asserting that the operators are all found
properly.